chore(dep): separate drawer into Editor folder
Use the Editor folder to build into editor specific dll
